Migrate project to use Gradle
