--- AndroidManifest.xml
@@ -1,36 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="uk.co.hgreen.hexclockwallpaper"
-    android:versionCode="3"
-    android:versionName="1.0.2" >
-    
-    <uses-sdk 
-        android:minSdkVersion="7" 
-        android:targetSdkVersion="20"/>
-    
-    <uses-feature
-        android:name="android.software.live_wallpaper"
-        android:required="true" >
-    </uses-feature>
-
-    <application
-        android:icon="@drawable/ic_launcher"
-        android:label="@string/app_name" 
-        android:allowBackup="true">
-        <service
-            android:name="HexclockWallpaperService"
-            android:enabled="true"
-            android:label="@string/label"
-            android:permission="android.permission.BIND_WALLPAPER" >
-            <intent-filter>
-                <action android:name="android.service.wallpaper.WallpaperService" >
-                </action>
-            </intent-filter>
-
-            <meta-data
-                android:name="android.service.wallpaper"
-                android:resource="@xml/hexclockwallpaper" >
-            </meta-data>
-        </service>
-    </application>
-
-</manifest>
