--- README.md
@@ -1,6 +1,0 @@
-# Hexclock Wallpaper
-
-A live wallpaper showing the current time in hexadecimal colour format.
-
-Based on Hexclock by Jacopo Colò:  
-http://www.jacopocolo.com/hexclock/
