+++ app/src/main/res/xml/hexclockwallpaper.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<wallpaper 
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  android:thumbnail="@mipmap/ic_launcher"
+  android:description="@string/wallpaper_description"/>
