--- res/xml/hexclockwallpaper.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<wallpaper 
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  android:thumbnail="@drawable/ic_launcher"
-  android:description="@string/wallpaper_description"/>
